+++ python/paddle/fluid/tests/unittests/mlu/test_conv2d_transposed_op_mlu.py
@@ -499,21 +499,21 @@ class TestConv2DTransposeAPI(unittest.TestCase):
         data2 = fluid.layers.data(
             name='data2', shape=[5, 5, 3], dtype='float32'
         )
-        out1 = fluid.layers.conv2d_transpose(
+        out1 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NCHW',
         )
-        out2 = fluid.layers.conv2d_transpose(
+        out2 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NHWC',
         )
-        out3 = fluid.layers.conv2d_transpose(
+        out3 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -521,7 +521,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [1, 1], [1, 1], [0, 0]],
             data_format='NHWC',
         )
-        out4 = fluid.layers.conv2d_transpose(
+        out4 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=3,
             num_filters=6,
@@ -529,7 +529,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [0, 0], [2, 1], [0, 0]],
             data_format='NCHW',
         )
-        out5 = fluid.layers.conv2d_transpose(
+        out5 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -537,7 +537,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='SAME',
             data_format='NCHW',
         )
-        out6 = fluid.layers.conv2d_transpose(
+        out6 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -545,7 +545,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='VALID',
             data_format='NHWC',
         )
-        out7 = fluid.layers.conv2d_transpose(
+        out7 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -586,7 +586,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         data = fluid.layers.data(name='data', shape=[3, 5, 5], dtype="float32")
 
         def attr_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -597,7 +597,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_data_format)
 
         def attr_padding_str():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -608,7 +608,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_str)
 
         def attr_padding_list():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -619,7 +619,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_list)
 
         def attr_padding_with_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -635,14 +635,14 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         )
 
         def error_input_size():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=error_input, groups=1, num_filters=6, filter_size=3
             )
 
         self.assertRaises(ValueError, error_input_size)
 
         def error_groups():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=0,
                 num_filters=6,
