+++ python/paddle/fluid/tests/unittests/npu/test_conv2d_transpose_op_npu.py
@@ -435,21 +435,21 @@ class TestConv2DTransposeAPI(unittest.TestCase):
         data2 = fluid.layers.data(
             name='data2', shape=[5, 5, 3], dtype='float32'
         )
-        out1 = fluid.layers.conv2d_transpose(
+        out1 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NCHW',
         )
-        out2 = fluid.layers.conv2d_transpose(
+        out2 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NHWC',
         )
-        out3 = fluid.layers.conv2d_transpose(
+        out3 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -457,7 +457,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [1, 1], [1, 1], [0, 0]],
             data_format='NHWC',
         )
-        out4 = fluid.layers.conv2d_transpose(
+        out4 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=3,
             num_filters=6,
@@ -465,7 +465,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [0, 0], [2, 1], [0, 0]],
             data_format='NCHW',
         )
-        out5 = fluid.layers.conv2d_transpose(
+        out5 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -473,7 +473,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='SAME',
             data_format='NCHW',
         )
-        out6 = fluid.layers.conv2d_transpose(
+        out6 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -481,7 +481,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='VALID',
             data_format='NHWC',
         )
-        out7 = fluid.layers.conv2d_transpose(
+        out7 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
