+++ python/paddle/fluid/tests/unittests/test_conv2d_transpose_op.py
@@ -835,21 +835,21 @@ class TestConv2DTransposeAPI(unittest.TestCase):
         data2 = fluid.layers.data(
             name='data2', shape=[5, 5, 3], dtype='float32'
         )
-        out1 = fluid.layers.conv2d_transpose(
+        out1 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NCHW',
         )
-        out2 = fluid.layers.conv2d_transpose(
+        out2 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NHWC',
         )
-        out3 = fluid.layers.conv2d_transpose(
+        out3 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -857,7 +857,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [1, 1], [1, 1], [0, 0]],
             data_format='NHWC',
         )
-        out4 = fluid.layers.conv2d_transpose(
+        out4 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=3,
             num_filters=6,
@@ -865,7 +865,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [0, 0], [2, 1], [0, 0]],
             data_format='NCHW',
         )
-        out5 = fluid.layers.conv2d_transpose(
+        out5 = paddle.static.nn.conv2d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -873,7 +873,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='SAME',
             data_format='NCHW',
         )
-        out6 = fluid.layers.conv2d_transpose(
+        out6 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -881,7 +881,7 @@ class TestConv2DTransposeAPI(unittest.TestCase):
             padding='VALID',
             data_format='NHWC',
         )
-        out7 = fluid.layers.conv2d_transpose(
+        out7 = paddle.static.nn.conv2d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -919,7 +919,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         data = fluid.layers.data(name='data', shape=[3, 5, 5], dtype="float32")
 
         def attr_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -930,7 +930,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_data_format)
 
         def attr_padding_str():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -941,7 +941,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_str)
 
         def attr_padding_list():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -952,7 +952,7 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_list)
 
         def attr_padding_with_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -968,14 +968,14 @@ class TestConv2DTransposeOpException(unittest.TestCase):
         )
 
         def error_input_size():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=error_input, groups=1, num_filters=6, filter_size=3
             )
 
         self.assertRaises(ValueError, error_input_size)
 
         def error_groups():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=0,
                 num_filters=6,
@@ -1064,7 +1064,7 @@ class TestTensorOutputSize3(TestTensorOutputSize1):
     def call_func(self, x):
         w_var = paddle.randn((3, 6, 3, 3), dtype='float32')
         output_size = paddle.assign([17])
-        out = paddle.fluid.layers.conv2d_transpose(
+        out = paddle.static.nn.conv2d_transpose(
             x, num_filters=6, output_size=output_size, filter_size=3, stride=2
         )
         return out
@@ -1076,7 +1076,7 @@ class TestTensorOutputSize4(TestTensorOutputSize1):
 
     def call_func(self, x):
         output_size = [17, paddle.assign([17])]
-        out = paddle.fluid.layers.conv2d_transpose(
+        out = paddle.static.nn.conv2d_transpose(
             x, num_filters=6, output_size=output_size, filter_size=3, stride=2
         )
         return out
