+++ python/paddle/fluid/tests/unittests/test_conv3d_transpose_part2_op.py
@@ -15,6 +15,7 @@
 import unittest
 import numpy as np
 
+import paddle
 import paddle.fluid.core as core
 import paddle.fluid as fluid
 from test_conv3d_transpose_op import TestConv3DTransposeOp
@@ -91,21 +92,21 @@ class TestConv3DTransposeAPI(unittest.TestCase):
             name='data2', shape=[5, 5, 5, 3], dtype='float32'
         )
 
-        out1 = fluid.layers.conv3d_transpose(
+        out1 = paddle.static.nn.conv3d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NCDHW',
         )
-        out2 = fluid.layers.conv3d_transpose(
+        out2 = paddle.static.nn.conv3d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
             filter_size=3,
             data_format='NDHWC',
         )
-        out3 = fluid.layers.conv3d_transpose(
+        out3 = paddle.static.nn.conv3d_transpose(
             input=data1,
             groups=1,
             num_filters=6,
@@ -113,7 +114,7 @@ class TestConv3DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [0, 0], [1, 1], [0, 0], [1, 1]],
             data_format='NCDHW',
         )
-        out4 = fluid.layers.conv3d_transpose(
+        out4 = paddle.static.nn.conv3d_transpose(
             input=data2,
             groups=3,
             num_filters=6,
@@ -121,7 +122,7 @@ class TestConv3DTransposeAPI(unittest.TestCase):
             padding=[[0, 0], [0, 0], [1, 1], [1, 2], [0, 0]],
             data_format='NDHWC',
         )
-        out5 = fluid.layers.conv3d_transpose(
+        out5 = paddle.static.nn.conv3d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -129,7 +130,7 @@ class TestConv3DTransposeAPI(unittest.TestCase):
             padding='SAME',
             data_format='NCDHW',
         )
-        out6 = fluid.layers.conv3d_transpose(
+        out6 = paddle.static.nn.conv3d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -137,7 +138,7 @@ class TestConv3DTransposeAPI(unittest.TestCase):
             padding='VALID',
             data_format='NDHWC',
         )
-        out7 = fluid.layers.conv3d_transpose(
+        out7 = paddle.static.nn.conv3d_transpose(
             input=data2,
             groups=1,
             num_filters=6,
@@ -177,7 +178,7 @@ class TestConv3DTransposeOpException(unittest.TestCase):
         )
 
         def attr_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -188,7 +189,7 @@ class TestConv3DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_data_format)
 
         def attr_padding_str():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -199,7 +200,7 @@ class TestConv3DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_str)
 
         def attr_padding_list():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
@@ -210,7 +211,7 @@ class TestConv3DTransposeOpException(unittest.TestCase):
         self.assertRaises(ValueError, attr_padding_list)
 
         def attr_padding_with_data_format():
-            out = fluid.layers.conv2d_transpose(
+            out = paddle.static.nn.conv2d_transpose(
                 input=data,
                 groups=1,
                 num_filters=6,
